refactor: Make quadprog, qpsolvers, and numpy optional dependencies
quadprog, qpsolvers, and numpy are now optional. They are grouped under the
`quadprog_projector` extra, which enables UPGrad and DualProj via
QuadprogProjector. numpy is also required by cagrad and nash_mtl, so it is
added to those extras too.

_WithOptionalDeps is moved from torchjd.aggregation._mixins to torchjd._mixins
so that it can be reused by QuadprogProjector in torchjd._linalg.

The standard install command is updated to `pip install "torchjd[quadprog_projector]"`
so that users following the README have access to UPGrad and DualProj out of the box.
