+++ src/torchjd/aggregation/_nash_mtl.py
@@ -9,7 +9,8 @@ import numpy as np
 import torch
 from torch import Tensor
 
-from torchjd.aggregation._mixins import Stateful, _NonDifferentiable, _WithOptionalDeps
+from torchjd._mixins import _WithOptionalDeps
+from torchjd.aggregation._mixins import Stateful, _NonDifferentiable
 
 from ._aggregator_bases import WeightedAggregator
 from ._weighting_bases import _MatrixWeighting
@@ -21,7 +22,7 @@ with contextlib.suppress(ImportError):
 
 # Non-differentiable: the cvxpy solver operates on numpy arrays, breaking the autograd graph.
 class _NashMTLWeighting(_WithOptionalDeps, _NonDifferentiable, Stateful, _MatrixWeighting):
-    _REQUIRED_DEPS = ["cvxpy", "ecos"]
+    _REQUIRED_DEPS = ["numpy", "cvxpy", "ecos"]
     _INSTALL_HINT = 'Install them with: pip install "torchjd[nash_mtl]"'
     """
     :class:`~torchjd.aggregation._mixins.Stateful`
